refactoring 85%
